Adding in Trilinos support into the "Merge" branch.
Implementing the Trilinos solvers with the same code 
base as the "parallel" branch. The code runs (on
my Linux workstation) in parallel with both Picard 
and JFNK solvers.


If the code is configured with Trilinos, the list of 
Trilinos libraries (which can vary greatly between
builds and versions) is automatically picked up 
using the logic in 
 libglimmer-trilinos/Makefile.Trilinos.export.in

I ran into an issue with Trilinos builds requiring
linking with CXX compiler instead of FC. I could
not get libtool to switch the linker based on
the conditional Trilinos dependency. I did not
want to force all builds to link with CXX. The 
best I could do was build a different executable,
so  "simple_glide_trilinos" is built for Trilinos 
builds, and "simple_glide" is built when trilinos
is not enabled. (Any libtool experts that can
fix this?) I disabled all other executables for
Trilinos builds. Let me know if verif, eis, 
executables should also have Trilinos versions.

More cleanup, porting, and documentation to come.
Andy 




git-svn-id: https://svn.berlios.de/svnroot/repos/glimmer-cism/glimmer-cism2/branches/glimmer-cism-lanl@2274 2cfc7cc4-626f-0410-8325-f6cb4dace9bf