Changed xml file for Trilinos options.
Renamed strat1.xml to trilinosOptions.xml. Not only is 
this a better name, it also is reformatted to make room 
for the nonlinear solvers options in the same file.

The formatting includes one level of nesting at the top:
<ParameterList>
  "strat1.xml" file contents
  <optional>  "input.xml" file contents for nonlinear solver
</ParameterList>

I added an error message and graceful exit() instead of 
core dump when there is a problem with the XML file.

I also added updated the content of the README file in 
libglimmer-trilinos directory.




git-svn-id: https://svn.berlios.de/svnroot/repos/glimmer-cism/glimmer-cism2/branches/glimmer-cism-lanl@2286 2cfc7cc4-626f-0410-8325-f6cb4dace9bf